Add files via upload
The K-Nearest Neighbors (KNN) project is a machine learning endeavor that applies the KNN algorithm to classify or predict data based on its proximity to neighboring data points. By measuring distances and considering K nearest data points, the model accurately predicts outcomes for new instances, making it a versatile and intuitive classification and regression tool. This project showcases the implementation and evaluation of KNN on various datasets, providing valuable insights into the algorithm's performance and applications in real-world scenarios.